Update readme + code cleanup
