+++ CogniteSdk.Types/Assets/AssetFilterDto.cs
@@ -37,6 +37,11 @@ namespace CogniteSdk.Assets
         public IEnumerable<Identity> AssetSubtreeIds { get; set; }
 
         /// <summary>
+        /// Only include assets that belong to these datasets.
+        /// </summary>
+        public IEnumerable<long> DataSetIds { get; set; }
+
+        /// <summary>
         /// Custom, application specific metadata. String key -> String value. Limits: Maximum length of key is 32
         /// bytes, value 512 bytes, up to 16 key-value pairs.
         /// </summary>
