+++ CogniteSdk.Types/Assets/AssetReadDto.cs
@@ -33,6 +33,11 @@ namespace CogniteSdk.Assets
         public string Description { get; set; }
 
         /// <summary>
+        /// Javascript friendly internal ID given to the object.
+        /// </summary>
+        public long? DataSetId { get; set; }
+
+        /// <summary>
         /// Custom, application specific metadata. String key -> String value
         /// </summary>
         public IDictionary<string, string> Metadata { get; set; }
